+++ lib/mayaUsd/fileio/shading/rfmShaderMap.h
